Changed integer to long ids in getExtensionById (fileExtensionsDao and fileExtensionsDaoImpl) and getCategotyById(fileCategoryDao and FilecategoryImpl)

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/FileCategoryDaoImpl.java
@@ -56,7 +56,7 @@ public class FileCategoryDaoImpl implements FileCategoryDao {
     }
 
     @Override
-    public Optional<String> getCategory(Integer categoryId) {
+    public Optional<String> getCategory(Long categoryId) {
         return jdbcTemplate.query("SELECT categoryName FROM file_categories WHERE categoryId = ?",new Object[]{categoryId}, FILE_CATEGORY_STRING_ROW_MAPPER).stream().findFirst();
     }
 }
